# fix: naive take on adding an index