--- a/packages/db/postgres/migrations/016-add-uploads-user-id-index.sql
+++ b/packages/db/postgres/migrations/016-add-uploads-user-id-index.sql
@@ -1,1 +1,0 @@
-CREATE INDEX IF NOT EXISTS upload_user_id_deleted_at_idx ON upload (user_id) WHERE deleted_at IS NULL;