user_id on uploads should probably be indexed...

completes #61