@drewdelano 2022-06-14
@vasco-santos This is the issue I'm looking at:
https://github.com/web3-storage/web3.storage/issues/61
https://github.com/web3-storage/web3.storage/blob/112b1788b8f226ae8d984132dedea90019441ea5/packages/db/index.js#L437

We seem to be seeing a steady increase over time which in my mind points to indexing.