@vasco-santos 2022-06-15
> That said, it might be worth running that query against staging to see how long it takes to return. It's entirely possible I'm wrong here.

We can definitely try it. Maybe the deleted is the key difference for users with a large number of uploads