@ribasushi 2022-06-15
@vasco-santos @drewdelano why not test all this ( EXPLAIN ANALYZE in hand ) on one of your replicas? The indices are all replicated and identical.