@vasco-santos 2022-06-15
> @vasco-santos @drewdelano why not test all this ( EXPLAIN ANALYZE in hand ) on one of your replicas? The indices are all replicated and identical.

Yes, good call @ribasushi . I was thinking about trying that before we move forward here. @drewdelano does not have access so must be me to try it