@vasco-santos 2022-06-15
It looks like this helps a lot! @drewdelano 

```
EXPLAIN ANALYZE 
SELECT * FROM upload
WHERE user_id = XXXXXXXXXXXXXXXXX
AND deleted_at IS NULL;
```

Before Index

```
 Bitmap Heap Scan on upload  (cost=6443.97..279801.88 rows=203188 width=279) (actual time=5260.205..68250.982 rows=187339 loops=1)
   Recheck Cond: (user_id = 'XXXXXXXXXXXXXXXXX'::bigint)
   Filter: (deleted_at IS NULL)
   Heap Blocks: exact=131036
   ->  Bitmap Index Scan on upload_user_id_source_cid_key  (cost=0.00..6433.81 rows=203799 width=0) (actual time=5228.591..5228.592 rows=187391 loops=1)
         Index Cond: (user_id = 'XXXXXXXXXXXXXXXXX'::bigint)
 Planning Time: 0.118 ms
 JIT:
   Functions: 4
   Options: Inlining false, Optimization false, Expressions true, Deforming true
   Timing: Generation 0.488 ms, Inlining 0.000 ms, Optimization 0.519 ms, Emission 3.815 ms, Total 4.822 ms
 Execution Time: 68301.216 ms
(12 rows)
```

After Index

```
 Index Scan using upload_user_id_deleted_at_idx on upload  (cost=0.09..216730.20 rows=202877 width=279) (actual time=0.071..525.612 rows=187339 loops=1)
   Index Cond: (user_id = 'XXXXXXXXXXXXXXXXX'::bigint)
 Planning Time: 3.819 ms
 JIT:
   Functions: 2
   Options: Inlining false, Optimization false, Expressions true, Deforming true
   Timing: Generation 0.194 ms, Inlining 0.000 ms, Optimization 0.000 ms, Emission 0.000 ms, Total 0.194 ms
 Execution Time: 564.808 ms
(8 rows)
```